+++ tools/configuration_common.py
