+++ tools/update_students_list.py
@@ -10,7 +10,7 @@ from bot.domain.constants import UniProgram
 from bot.domain.student.attribut.discord_user_id import DiscordUserId
 from bot.infra.constants import StudentMongoDbKey
 from constants import StudentCsvKey
-from tools.common import get_configuration, add_configuration_argument
+from tools.configuration_common import get_configuration, add_configuration_argument
 
 ARGUMENT_FILENAME_KEY = "csv_filename"
 STUDENTS_LIST_COLUMNS_TO_KEEP: List[str] = [
